+++ platforms/web/test-e2e/tests/home_test.ts
@@ -39,6 +39,52 @@ Scenario('Header button navigates to playlist screen', async ({ I }) => {
   I.see('The Daily Dweebs');
 });
 
+// @todo: this test requires the testconfig to have `layoutType: 'hero'` on the first shelf
+// eslint-disable-next-line codeceptjs/no-skipped-tests
+Scenario.skip('I can slide within the hero shelf', async ({ I }) => {
+  const isDesktop = await I.isDesktop();
+  const visibleTileLocator = locate({ css: `section[data-testid="shelf-${ShelfId.hero}"] div[data-testid="shelf-hero-metadata--visible"]` });
+
+  async function slide(swipeText: string, forward = true) {
+    if (isDesktop) {
+      I.click({ css: `button[aria-label="${forward ? 'Next' : 'Previous'} slide"]` });
+    } else {
+      forward ? await I.swipeLeft({ text: swipeText }) : await I.swipeRight({ text: swipeText });
+    }
+  }
+
+  await within(makeShelfXpath(ShelfId.hero), async () => {
+    I.see('Blender Channel');
+  });
+
+  await within(visibleTileLocator, function () {
+    I.dontSee('Spring');
+    I.dontSee('Spring is a 2019 animated');
+  });
+
+  await within(makeShelfXpath(ShelfId.hero), async () => {
+    await slide('Blender Channel');
+  });
+
+  await within(visibleTileLocator, function () {
+    I.waitForElement('text=Spring', 3);
+    I.see('Spring is a 2019 animated');
+  });
+
+  // Without this extra wait, the second slide action happens too fast after the first and even though the
+  // expected elements are present, the slide doesn't work. I think there must be a debounce check on the carousel.
+  I.wait(1);
+
+  await within(makeShelfXpath(ShelfId.hero), async () => {
+    await slide('Spring', false);
+  });
+
+  await within(visibleTileLocator, function () {
+    I.waitForElement('text="Blender Channel"', 3);
+    I.dontSee('Spring');
+  });
+});
+
 Scenario('I can slide within the featured shelf', async ({ I }) => {
   const isDesktop = await I.isDesktop();
   const visibleTilesLocator = locate({ css: `section[data-testid="shelf-${ShelfId.featured}"] .TileSlider--visible` });