+++ platforms/web/test-e2e/tests/live_channel_test.ts
@@ -5,7 +5,7 @@ import { testConfigs } from '@jwp/ott-testing/constants';
 
 import { ShelfId } from '#utils/constants';
 
-const programSelectedBackgroundColor = 'rgb(204, 204, 204)';
+const programSelectedBackgroundColor = 'rgb(255, 255, 255)';
 const programLiveBorder = '2px solid rgb(255, 255, 255)';
 
 const programBackgroundColor = 'rgba(255, 255, 255, 0.08)';
@@ -158,6 +158,7 @@ Scenario('I can select an upcoming program on the same channel', async ({ I }) =
 
 Scenario('I can select a previous program on the same channel and watch the video', async ({ I }) => {
   await I.openVideoCard('Channel 1');
+  waitForEpgAnimation(I);
   I.seeElement(channel1LiveProgramLocator);
   await isSelectedProgram(I, channel1LiveProgramLocator, 'channel 1', true);
 
@@ -184,6 +185,7 @@ Scenario('I can select a previous program on the same channel and watch the vide
 
 Scenario('I can select a program on another channel', async ({ I }) => {
   await I.openVideoCard('Channel 1');
+  waitForEpgAnimation(I);
   I.click(channel2Locator);
 
   waitForEpgAnimation(I);
@@ -268,7 +270,7 @@ Scenario('I can see the background image for Channel 4', async ({ I }) => {
 async function isSelectedProgram(I: CodeceptJS.I, locator: CodeceptJS.Locator, channel: string, isLive: boolean) {
   I.moveCursorTo('body', 0, 0); // This prevents accidentally triggering the hover state
 
-  await checkStyle(I, locator, {
+  await I.seeCssProperties(locator, {
     'background-color': programSelectedBackgroundColor,
     border: isLive ? programLiveBorder : programBorder,
   });
@@ -277,7 +279,7 @@ async function isSelectedProgram(I: CodeceptJS.I, locator: CodeceptJS.Locator, c
 }
 
 async function isLiveProgram(I: CodeceptJS.I, locator: CodeceptJS.Locator, channel: string) {
-  await checkStyle(I, locator, {
+  await I.seeCssProperties(locator, {
     'background-color': programBackgroundColor,
     border: programLiveBorder,
   });
@@ -286,16 +288,12 @@ async function isLiveProgram(I: CodeceptJS.I, locator: CodeceptJS.Locator, chann
 }
 
 async function isProgram(I: CodeceptJS.I, locator: CodeceptJS.Locator, channel: string) {
-  await checkStyle(I, locator, {
+  await I.seeCssProperties(locator, {
     'background-color': programBackgroundColor,
     border: programBorder,
   });
 
   I.say(`I see the program is not active nor selected on ${channel}`);
-}
-
-async function checkStyle(I: CodeceptJS.I, locator: CodeceptJS.LocatorOrString, styles: Record<string, string>) {
-  I.seeCssPropertiesOnElements(locator, styles);
 }
 
 function waitForEpgAnimation(I: CodeceptJS.I, sec: number = 1) {